+++ Assets/Scripts/UI/Inventory/InventoryUI.cs
@@ -48,7 +48,7 @@ public class InventoryUI : MonoBehaviour
     [Tooltip("This should only be on if you want to dynamically change UI positioning through the editor. Otherwise, leave this off for performance.")]
     public bool ShouldUpdateSlotPositions = false;
 
-    private CharacterInventoryAndInteraction CharacterInventory;
+    private EntityInventory CharacterInventory;
     private IList<InventorySlotUI> MainInventorySlots;
     private IList<InventorySlotUI> HotbarInventorySlots;
     private MouseSlotUI MouseInventorySlot;
